requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-tailwind==1.0.3
cryptography==44.0.0
Django==5.1.2
mysqlclient==2.2.1
django-allauth==65.0.2
django-cors-headers==4.5.0
django-crispy-forms==2.3
django-encrypted-model-fields==0.6.5
django-filter==24.3
django-ratelimit==4.1.0
django-redis==5.4.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==30.8.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
kombu==5.4.2
numpy==2.2.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
reportlab==4.2.5
requests==2.32.3
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
XlsxWriter==3.2.0
python-dotenv==1.0.1
django-select2==8.4.0